+++ app/assets/javascripts/main.js
@@ -44,3 +44,4 @@
 //= require bootstrap
 //= require osmlocator
 //= require bootstrap-switch
+//= require blueimp-gallery